[Truffle] Removed redundant casts.

+++ truffle/src/main/java/org/jruby/truffle/nodes/rubinius/ObjectPrimitiveNodes.java
@@ -94,7 +94,7 @@ public abstract class ObjectPrimitiveNodes {
                 return newId;
             }
 
-            return (long) id;
+            return id;
         }
 
         protected ReadHeadObjectFieldNode createReadObjectIDNode() {
